Integrity, truth, compulsory education, wealth, peace, tyranny - Learn to live free
Is it possible to be free and not live with integrity?  

Is there an essential connection between the truth and being free?  

It is the height of opression to assert my right to be free, while ignoring my fellow citizens's right to be free.  Can and should this really exist?  

When someone intervenes between a child and the child's parents asserting or complelling the right to control that child's education, what impact does that have on freedom for that child, those parents, and that society?  

It is hard to win freedom.  It is easy to surrender freedom.  Freedom is something tyrrants despise.  Tyrrany and freedom are opposites.  

The most common reasons that people surrender freedom is for personal peace and for affluence.  The person who asks you to surrender your freedom to enjoy peace will lead you to war.  The person who asks you to surrender your freedom so you can enjoy affluence will lead you to the death strangles of poverty.  That person does not care for your peace or for your wealth, other than the fact that for you to have those means that he does not and he will not have that.  He will have your peace and your wealth.